connecting to mongodb
There I connect this application to mongodb atlas. That's a cloud database this is a simple one. First I try to connect with mysql database but I don't found that it's cloud or not then I make a new account in mongodb and use it's free clustuer. It's hard to make connection but I mange it but there is a problem that its connect but unique key constraint is not work properly my data that I give It from Thender Strom vs-code extension It's became duplicate when I send It again. I need to solve it I will solve it as soon as possible. I think I make a stay because in my college my sendups are coming I need to prepare them on 14 January 2025 my first test is and now it's 12 January I have only two days left So if i make some leaves so never mind It's important.